common/reexec: move to internal/reexec
