+++ p2p/simulations/adapters/exec.go
@@ -34,7 +34,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common/reexec"
+	"github.com/ethereum/go-ethereum/internal/reexec"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/p2p"
