+++ p2p/simulations/adapters/types.go
@@ -25,8 +25,8 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/ethereum/go-ethereum/common/reexec"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/internal/reexec"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/p2p"
